requirements.txt
asgiref==3.8.1
certifi==2024.8.30
Django==4.2
django-jazzmin==3.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
pillow==10.4.0
sqlparse==0.5.1
